Implement Channel's Close method.
The close method has been converted into an operation that is queued on
the Channel's internal command channel; this makes the public method
thread safe. The closing of a channel is done in two steps. First, the
Channel is closed which makes all input/output impossible on it. Secondly,
the Channel's internal command channel is closed which effectively closes
the channel. A new state is added to represent a disposed channel.