Fix a bug where the CI param wasn't set correctly.
The CI parameter is set by the client to signal that the bind request
should be closed immediately by the server; it disables the chunking
support. The client sets this parameter based on response form the
second test request. If the response isn't flushed properly, the client
will infer that a buffering proxy is present on the connection which
isn't the case.